back/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
pyngrok
nest_asyncio